Allow requestingRoutes to be a valid (but ignored) state when tapping… (#73)
… on origin/destination update.

Some routes take a while to request and parse. This is the simplest
possible fix, but ignore the tap on the rendered start/end point links
on directions preview while still in .requestingRoutes state. A better
fix would involve some UX to show progress indicator or something along
those lines.